New API for constant B-Spline interpolation
